% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatclass.R
\name{extat_class}
\alias{extat_class}
\title{Classification statistics and table}
\usage{
extat_class(model, dep.var, prob_cut)
}
\arguments{
\item{model}{The regression model that was stored prior}

\item{dep.var}{The observed dependent variable (with data frame as prefix, "df$dep.var")}

\item{prob_cut}{cut-off point at which the predicted probabilities should be coded binary (0,1). Usually 0.5 is used to indicate >0.5 as 1 and <0.5 as 0}
}
\value{
Different class-values and a list of them
}
\description{
Produces an extended classification table and statistics given a binary response model.
}
